LayoutTests/platform/ios-sim/fast/forms/file/input-file-value-expected.txt

CONSOLE MESSAGE: line 35: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(files)')
This tests the value attribute in file input forms

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
